test/sdi: fixed x86_64 built. Removed remaining compiler warnings.
git-svn-id: https://svn.aros.org/svn/aros/trunk@49545 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ test/sdi/examples/hooks/example_hookuse.c
@@ -74,7 +74,7 @@ int main(void)
 
       ret = CallHook(&HelloHook, "portable Hook");
 
-      printf("%ld\n", ret);
+      printf("%ld\n", (long int)ret);
 
       DROPINTERFACE(IUtility);
     }
